refactor: create dedicated module and scripts for experiments
This commit refactor the program to be able to run more easily a various experiments by creating a dedicated experimenter. The experimenter load an experiment script and run the test automatically. This will allow to not have to modify most of the code whenever a new experiment needs to be added. The commit also introduce the notion of drivers: The driver run automatically several external programs like SDN controllers or the Fuzzer and so on.

+ Add some shell scripts to automatically install/uninstall onos
+ Add onos driver
+ Add experiment "onos_2node_1ping" and one criteria
+ Add experimenter module

~ Bump version number to 0.1.0
~ Change logging format: now a single log file is created under .rdfl_exp instead of an individual log file per experiment
~ Refactor to account for the addition of the experimenter and the drivers

~ Fix an issue in rule.py#convert_to_fuzzer_actions where an error would be raised if the generation of a single action was requested

- Remove old implementation of script.py