fix: allow iteration to continue even if there is no rules.
This commit allows the experiments to continue even if there no rules where generated due to various issues (i.e. Not enough datapoint for SMOTE, no error detected, and so on.). This also involed changes in stats.py and rdfl_exp-summarize. The latter changes will be pushed once the script is generalized to not include personal information.