refactor!: merge executable of figsdn and figsdn-report
This commit merges figsdn-report and figsdn executable file, and by doing so, also changes the cli commands. The cli commands now takes two main commands: 'experiments' and 'nodes'. Using 'nodes' allow to manage the registered nodes for all modules and 'experiments' deals with the experiments.
The experiments command has two major sub-commands: 'run' and 'report', 'run' allows to launch the app and the test of a scenario and 'report' reports the results of a previous experiment.

+ Add a main.py file to figsdn, which is the entry point for the cli
~ Refactor figsdn.app.main to fit the new cli
~ Refactor figsdn.app.setup to fit the new cli
~ Refactor figsdn.report.main to fit the new cli
- Remove figsdn-report executable

BREAKING CHANGE: Complete change of cli API