fix: resolve issue in figsdn.app.experimenter with budget calculation
Resolve an issue where the budget calculation, when using the confidence mode, was not correctly attributed. This was due to the context no being declared as a "global" in the "run" function, and thus not being updated accross all function of figsdn.app.main.
The patch consist in both adding "global _context" in figsdn.app.main#calculate_budget and using in the same function the "target" argument for both methods