fix: resolve an issue with 'has_error' column data type
This commit addresses a database-related issue involving the 'has_error' column in the log table.
Previously, this column used the 'BIT(1)' data type, which caused problems due to MySQL's handling of data storage. Furthermore, using 'BIT(1)' instead of 'TINYINT(1)' offered little to no advantages.
As a result, the data type was switched to 'TINYINT(1)' to resolve this issue and improve data management.